reduce lease time to 2minutes
tinc on restart will generate a new mac address, which means that when a box
reboots, the current lease will be blocked by a different mac address.
By setting the lease to 2 minutes, we can deal with a box that restarts and
will be able to re-acquire its ip address on a different mac address as its old
lease should have now expired.